Clamp alpha, histogram dtype, and assorted fixes
Preserve alpha in clamp_normalized and return float32; normalize histogram outputs to numeric dtypes and rename cumulative_histogram module. Apply various typing/casting fixes and small bug fixes across the raster package: add type:ignore annotations for OpenCV/torch returns, ensure kernel usage for morphology, add sanity asserts (preprocess pipeline), fix channel name handling, normalize KMeans/cv2 calls, and strengthen depth estimator checks and normalization. Also update tests to assert module specs and suppress type issues. These changes improve type-safety, correctness for RGBA handling, and consistency of numeric types.

+++ src/dreamstack/raster/analysis/histogram/histogram_rgb.py
@@ -34,7 +34,19 @@ def histogram_rgb(
     tuple
         (blue_hist, green_hist, red_hist)
     """
-    b_hist = cv2.calcHist([image], [0], mask, [bins], [0, 256]).flatten()
-    g_hist = cv2.calcHist([image], [1], mask, [bins], [0, 256]).flatten()
-    r_hist = cv2.calcHist([image], [2], mask, [bins], [0, 256]).flatten()
+    b_hist = (
+        cv2.calcHist([image], [0], mask, [bins], [0, 256])
+        .flatten()
+        .astype(np.float64)
+    )
+    g_hist = (
+        cv2.calcHist([image], [1], mask, [bins], [0, 256])
+        .flatten()
+        .astype(np.float64)
+    )
+    r_hist = (
+        cv2.calcHist([image], [2], mask, [bins], [0, 256])
+        .flatten()
+        .astype(np.float64)
+    )
     return (b_hist, g_hist, r_hist)
